refactor: initial flyway and change map struct

+++ src/main/java/com/ecommerce/sshop/repository/role/IRoleRepository.java
@@ -6,7 +6,7 @@ import com.ecommerce.sshop.model.role.Role;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IRoleRepository extends JpaRepository<Role, Long> {
+public interface IRoleRepository extends JpaRepository<Role, Short> {
     Optional<Role> findByName(String roleName);
 
 }
